Update read seed
