--- android/app/src/main/kotlin/com/bitmark/autonomywallet/FileStorageHelper.kt
@@ -1,17 +1,0 @@
-package com.bitmark.autonomywallet
-
-import android.content.Context
-import java.io.File
-
-class FileStorageHelper(private val context: Context) {
-
-    fun listFiles(): List<String> {
-        val filesDir = context.filesDir
-        return filesDir.list()?.toList() ?: emptyList()
-    }
-
-    fun readOnFilesDir(fileName: String): ByteArray {
-        val file = File(context.filesDir, fileName)
-        return file.readBytes()
-    }
-}
