+++ android/app/src/main/kotlin/com/bitmark/autonomywallet/MainActivity.kt
@@ -6,321 +6,45 @@
 
 package com.bitmark.autonomy_flutter
 
-import android.app.KeyguardManager
-import android.content.Context
-import android.content.Intent
-import android.content.SharedPreferences
 import android.os.Build
 import android.util.Log
-import android.view.WindowManager.LayoutParams
-import androidx.biometric.BiometricManager
-import com.bitmark.autonomywallet.FileStorageHelper
-import com.bitmark.libauk.model.Seed
-import com.bitmark.libauk.util.fromJson
-import com.bitmark.libauk.util.newGsonInstance
-import com.google.android.gms.auth.blockstore.Blockstore
 import com.google.android.gms.auth.blockstore.BlockstoreClient
 import com.google.android.gms.auth.blockstore.BlockstoreClient.DEFAULT_BYTES_DATA_KEY
 import com.google.android.gms.auth.blockstore.RetrieveBytesRequest
-import com.google.gson.Gson
-import com.google.gson.reflect.TypeToken
 import io.flutter.embedding.android.FlutterFragmentActivity
 import io.flutter.embedding.engine.FlutterEngine
-import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
-import java.time.format.DateTimeParseException
-import java.util.concurrent.TimeUnit
+import kotlinx.serialization.SerialName
 
 class MainActivity : FlutterFragmentActivity() {
     companion object {
-        var isAuthenticate: Boolean = false
-        private const val CHANNEL = "migration_util"
-        private val secureScreenChannel = "secure_screen_channel"
         private val systemChanel = "com.feralfile.wallet"
-        private var lastAuthTime: Long = 0
-        private val authenticationTimeout = TimeUnit.MINUTES.toMillis(3)
         private lateinit var client: BlockstoreClient
     }
-
-    var flutterSharedPreferences: SharedPreferences? = null
-
-    private val backupDartPlugin: BackupDartPlugin = BackupDartPlugin()
-
-    //DONT REMOVE; We will bring back this code when we need to verify the signature
-    // check if the signature is valid
-//    private fun isSignatureValid(
-//        context: Context,
-//        expectedSignatureHash: String
-//    ): Boolean {
-//        try {
-//            val packageInfo = context.packageManager.getPackageInfo(
-//                context.packageName,
-//                PackageManager.GET_SIGNING_CERTIFICATES
-//            )
-//            val signatures = packageInfo.signingInfo.apkContentsSigners
-//            val md = MessageDigest.getInstance("SHA-256")
-//            for (signature in signatures) {
-//                md.update(signature.toByteArray())
-//                val currentSignatureHash = Base64.encodeToString(md.digest(), Base64.NO_WRAP)
-//                if (currentSignatureHash == expectedSignatureHash) {
-//                    return true
-//                }
-//            }
-//        } catch (e: Exception) {
-//            Log.e("Signature", e.message.toString())
-//        }
-//        return false
-//    }
 
     override fun configureFlutterEngine(flutterEngine: FlutterEngine) {
         super.configureFlutterEngine(flutterEngine)
         FileLogger.init(applicationContext)
         // verity signing certificate
 
-        // DONT REMOVE, we will bring back this code when we need to verify the signature
-//        checkSecurity()
-        client = Blockstore.getClient(applicationContext)
-        MethodChannel(
-            flutterEngine.dartExecutor.binaryMessenger,
-            CHANNEL
-        ).setMethodCallHandler { call, result ->
-            if (call.method == "getExistingUuids") {
-                val uuids = getExistingUuids()
-                result.success(uuids)
-            } else {
-                result.notImplemented()
-            }
-        }
-
-//        backupDartPlugin.createChannels(flutterEngine, applicationContext)
-
-        MethodChannel(
-            flutterEngine.dartExecutor.binaryMessenger,
-            secureScreenChannel
-        ).setMethodCallHandler { call, result ->
-            if (call.method == "setSecureFlag") {
-                val secure = call.argument<Boolean>("secure") ?: false
-                setSecureFlag(secure)
-                result.success(null)
-            } else {
-                result.notImplemented()
-            }
-        }
 
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
             window.setHideOverlayWindows(true)
         }
 
-        val fileStorageHelper = FileStorageHelper(this)
         val systemChannel = MethodChannel(flutterEngine.dartExecutor.binaryMessenger, systemChanel)
         systemChannel.setMethodCallHandler { call, result ->
             when (call.method) {
-                "exportMnemonicForAllPersonaUUIDs" -> exportMnemonicForAllPersonaUUIDs(call, result)
+                "exportMnemonicForAllPersonaUUIDs" -> exportMnemonicForAllPersonaUUIDs(result)
             }
         }
     }
 
-    // DONT REMOVE, we will bring back this code when we need to verify the signature
-    // detect frida
-//    private fun detectFrida(): Boolean {
-//        return detectFridaPort() || detectFridaMem()
-//    }
 
-    // DONT REMOVE, we will bring back this code when we need to verify the signature
-    // check security
-//    private fun checkSecurity() {
-//        val handler = Handler(Looper.getMainLooper())
-//
-//        handler.postDelayed({
-//            if (detectFrida()) {
-//                captureMessage("[Security check] Reverse engineering tool detected")
-//                finish()
-//            }
-//            val isSignatureValid = isSignatureValid(this, BuildConfig.SIGNATURE_HASH)
-//            if (!isSignatureValid) {
-//                captureMessage("[Security check] Invalid signature detected")
-//                Toast.makeText(this, "Invalid signature", Toast.LENGTH_SHORT).show()
-//                finish()
-//            }
-//
-//            // Detect rooted devices
-//            // Create a RootBeer instance
-//            val rootBeer = RootBeer(this)
-//            if (rootBeer.isRooted) {
-//                captureMessage("[Security check] Rooted device detected")
-//                Toast.makeText(this, "This app cannot be used on rooted devices.", Toast.LENGTH_SHORT)
-//                    .show()
-//                finish() // Close the app
-//            }
-//
-//            // debugger detection
-//            val hasTracerPid = hasTracerPid()
-//            if (BuildConfig.ENABLE_DEBUGGER_DETECTION && hasTracerPid) {
-//                captureMessage("[Security check] Debugger detected")
-//                Toast.makeText(
-//                    this,
-//                    "Debugging detected. Please try again without any debugging tools.",
-//                    Toast.LENGTH_SHORT
-//                )
-//                    .show()
-//                finish()
-//            }
-//        }, 5000L)
-//    }
-
-    // DONT REMOVE, we will bring back this code when we need to verify the signature
-//    private fun captureMessage(message: String) {
-//        try {
-//            Sentry.captureMessage(message)
-//        } catch (e: Exception) {
-//            e.printStackTrace()
-//        }
-//    }
-
-
-    // DONT REMOVE, we will bring back this code when we need to verify the signature
-    // detect frida
-//    private fun detectFridaPort(): Boolean {
-//        return try {
-//            val socket = Socket()
-//            socket.connect(InetSocketAddress("127.0.0.1", 27042), 1000)
-//            socket.close()
-//            true
-//        } catch (e: Exception) {
-//            false
-//        }
-//    }
-
-    // DONT REMOVE, we will bring back this code when we need to verify the signature
-//    private fun detectFridaMem(): Boolean {
-//        try {
-//            val mapsFile = BufferedReader(FileReader("/proc/self/maps"))
-//            var isFridaDetected = false
-//
-//            while (true) {
-//                val line = mapsFile.readLine() ?: break
-//
-//                if (line.contains("frida")) {
-//                    isFridaDetected = true
-//                    break
-//                }
-//            }
-//
-//            mapsFile.close()
-//            return isFridaDetected
-//        } catch (e: Exception) {
-//            e.printStackTrace()
-//            return false
-//        }
-//    }
-
-    private fun getExistingUuids(): String {
-        val sharedPreferences = this.getSharedPreferences(
-            BuildConfig.APPLICATION_ID, Context.MODE_PRIVATE
-        )
-        return sharedPreferences.getString("persona_uuids", "") ?: ""
-    }
-
-    private fun setSecureFlag(secure: Boolean) {
-        if (secure) {
-            window.addFlags(LayoutParams.FLAG_SECURE)
-        } else {
-            window.clearFlags(LayoutParams.FLAG_SECURE)
-        }
-    }
-
-    override fun onResume() {
-        super.onResume()
-
-        val sharedPreferences = flutterSharedPreferences ?: this.getSharedPreferences(
-            "FlutterSharedPreferences",
-            Context.MODE_PRIVATE
-        )
-        val isEnabled = sharedPreferences.getBoolean("flutter.device_passcode", false)
-
-//        backupDartPlugin.getJWT { jsonString ->
-//            if (jsonString.isNullOrEmpty()) {
-//                Log.e("getJWT", "JWT retrieval failed or returned null")
-//                return@getJWT
-//            }
-//            // decode jsonString
-//            val type = object : TypeToken<Map<String, Any>>() {}.type
-//            val map: Map<String, Any> = Gson().fromJson(jsonString, type)
-//            val refreshTokenExpiredAtString = map["refresh_expire_at"] as? String ?: run {
-//                Log.e("getJWT", "Missing 'refresh_expire_at' in JWT")
-//                return@getJWT
-//            }
-//
-//            val refreshTokenExpiredAt = try {
-//                java.time.Instant.parse(refreshTokenExpiredAtString)
-//            } catch (e: DateTimeParseException) {
-//                Log.e("getJWT", "Invalid date format: $refreshTokenExpiredAtString")
-//                return@getJWT
-//            }
-//            val isExpired = refreshTokenExpiredAt.isBefore(java.time.Instant.now())
-//            if (isExpired) {
-//                updateAuthenticationTime()
-//                return@getJWT
-//            }
-//            if (isEnabled && !isAuthenticate && needsReAuthentication()) {
-//                val biometricManager = BiometricManager.from(this)
-//                val keyguardManager = getSystemService(KEYGUARD_SERVICE) as KeyguardManager
-//                if (biometricManager.canAuthenticate(BiometricManager.Authenticators.BIOMETRIC_STRONG)
-//                    == BiometricManager.BIOMETRIC_SUCCESS || keyguardManager.isDeviceSecure
-//                ) {
-//                    val intent = Intent(this@MainActivity, AuthenticatorActivity::class.java)
-//                    updateAuthenticationTime()
-//                    startActivity(intent)
-//                }
-//            }
-//        }
-    }
-
-    private fun updateAuthenticationTime() {
-        lastAuthTime = System.currentTimeMillis()
-    }
-
-    private fun needsReAuthentication(): Boolean {
-        val currentTime = System.currentTimeMillis()
-        return (currentTime - lastAuthTime) > authenticationTimeout
-    }
-
-    override fun onPause() {
-        super.onPause()
-        isAuthenticate = false
-    }
-
-    // DONT REMOVE, we will bring back this code when we need to verify the signature
-//    private fun hasTracerPid(): Boolean {
-//        val tracerpid = "TracerPid"
-//        try {
-//            val file = File("/proc/self/status")
-//            val lines = file.readLines()
-//            for (line in lines) {
-//                if (line.length > tracerpid.length) {
-//                    if (line.substring(0, tracerpid.length).equals(tracerpid, ignoreCase = true)) {
-//                        val pid = line.substring(tracerpid.length + 1).trim().toInt()
-//                        if (pid > 0) {
-//                            return true
-//                        }
-//                        break
-//                    }
-//                }
-//            }
-//        } catch (exception: Exception) {
-//            println(exception)
-//        }
-//        return false
-//    }
-
-    private fun exportMnemonicForAllPersonaUUIDs(call: MethodCall, result: MethodChannel.Result) {
+    private fun exportMnemonicForAllPersonaUUIDs(result: MethodChannel.Result) {
         val retrieveBytesRequestBuilder = RetrieveBytesRequest.Builder()
             .setRetrieveAll(true)
-        
+
         client.retrieveBytes(retrieveBytesRequestBuilder.build())
             .addOnSuccessListener { bytes ->
                 try {
@@ -331,12 +55,21 @@ class MainActivity : FlutterFragmentActivity() {
                         defaultBytesData?.bytes?.toString(Charsets.UTF_8) ?: ""
                     )
 
-                    val mnemonics = data.accounts.map { it.mnemonic }
-                    result.success(mnemonics)
-                    
+                    val mnemonicsMap =
+                        mutableMapOf<String, List<String>>() // Map<String, List<String>>
+
+                    data.accounts.forEach { account ->
+                        val uuid = account.uuid
+                        val mnemonic = account.mnemonic.split(" ")
+                        val passphrase = account.passphrase ?: ""
+
+                        mnemonicsMap[uuid] = listOf(passphrase) + mnemonic
+                    }
+                    result.success(mnemonicsMap)
                 } catch (e: Exception) {
+                    e.printStackTrace()
                     // No accounts found
-                    result.success(emptyList<String>())
+                    result.success(emptyMap<String, List<String>>())
                 }
             }
             .addOnFailureListener { e ->
@@ -345,3 +78,21 @@ class MainActivity : FlutterFragmentActivity() {
             }
     }
 }
+
+@Serializable
+data class BackupData(
+    @SerialName("accounts")
+    val accounts: List<BackupAccount>
+)
+
+@Serializable
+data class BackupAccount(
+    @SerialName("uuid")
+    val uuid: String,
+    @SerialName("mnemonic")
+    val mnemonic: String,
+    @SerialName("passphrase")
+    val passphrase: String?,
+    @SerialName("name")
+    val name: String,
+)
